Use TypeChatTranslator instead of custom TextClassiferTranslator
